requirements.txt
astunparse==1.6.3
async-timeout==3.0.1
discord-py-interactions==3.0.2
discord-py-slash-command==3.0.3
discord-together==1.2.6
discord.py==1.7.3
# ffmpeg==1.4
PyNaCl==1.5.0
uritemplate==3.0.1
# urllib3==1.26.13
# youtube-dl==2021.12.17
# yt-dlp==2022.6.29
asyncio
pyunsplash
asyncpraw
pypexels

